Able to add elements for participants.
In order to publish the modifications of the lists in guests but also in hosts,
The following protocol is needed:

```js
const newList = GuestList.copy(this._list);
newList.add(id); // or remove
this._updateList(newList);
```

In other words, DO NOT update the list directly because only _updateList() will
call _informChanges() which is in charge of emitting the proper update events.
This limitation should be removed by adding an update() method to GuestList to
transform the current state into the new one without the need of reassigning a
new list and adding events to the model. Then, any intent of updating the list
will result in the modifications being notified.

The GuestList should be a GuestLog with all the participant operations (enters
and exits). The log not only guarantee enter order but also exit.

If the log is not implemented, at least we should change nullifying a
participant when it exits with marking it as absent.

Communication of events have been fixed. The final design of how to add a
participant is not final yet. Perhaps it is better to create a new entity for
participants and add the contents of the template inside.