Added participant monitorization.
It still remains allowing customization of the templating process.

It would be better if all related to monitoring local participant updates would
be isolated in its own module or class.

Fixed a silly error: I was monitoring all the participants in the scene in each
peer and broadcasting the modifications happening in all of them. This was
causing a cascade reaction after applying the updates coming from another peer.
Only the owner of an especific participant must broadcast its updates.

The "shared" component has been renamed to "myself" and it only serves to
annotate which components must be broadcasted to peers. This should be part of
the customization process but, by default, it shares the "rotation" only.