Adding key updates to force global sync from time to time.
These key updates send the current values of the shared components without
waiting for a component to change. They happen each 60 checks. Perhaps it should
be configurable.

Thinking of changing the API. Consider using `isLocal` instead of `isMe` and
`mixonlocal` instead of `onmyself`. We need a solid term to refer to ourselves.