Participation when role is host tested.
It would remain to test order of the events when exiting before entering is
complete. It is needed to enforce that, if there is an exit event, it is sure
there've been an enter event before.

Some conditions are still needed for ensuring participantmessage and
participantstream events always happen after enterparticipant has happened.